BUG: use fabs not abs for gcc compatibility
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=1055
git-svn-id: http://svn.slicer.org/Slicer4/trunk@1055 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Base/Logic/vtkImageSlicePaint.cxx
@@ -58,7 +58,7 @@ void vtkImageSlicePaintExtract(vtkImageSlicePaint *self, T *ptr)
   double deltaBottomRow[3];
   double deltaRightColumn[3];
   double deltaLeftColumn[3];
-  int maxRowDelta = 0, maxColumnDelta = 0;
+  double maxRowDelta = 0, maxColumnDelta = 0;
   int maxRowIndex, maxColumnIndex;
 
   // first get the width and height of the extracted region
